+++ packages/ui/src/List/List/ListItemAction.tsx
@@ -1,0 +1,19 @@
+import React from "react";
+
+export type ListActionsProps = {
+    children: React.ReactNode;
+};
+
+/**
+ * Used to contain the actions inside ListItemMate component.
+ * @param props
+ * @returns {*}
+ * @constructor
+ */
+export const ListActions = (props: ListActionsProps) => {
+    return (
+        <span {...props} className={"webiny-list-actions"}>
+            {props.children}
+        </span>
+    );
+};
