--- packages/ui/src/List/README.md
@@ -1,22 +1,0 @@
-# List
-
-### Design
-
-<a href="https://material.io/design/components/lists.html" target="_blank">https://material.io/design/components/lists.html</a>
-
-### Description
-
-Use `List` component to display data and offer additional actions if needed.
-
-### Import
-
-```js
-import {
-  List,
-  ListItem,
-  ListItemText,
-  ListItemTextSecondary,
-  ListItemGraphic,
-  ListItemMeta
-} from "@webiny/ui/List";
-```
