--- packages/ui/src/List/icons/index.ts
@@ -1,4 +1,0 @@
-import refreshIcon from "./baseline-autorenew-24px.svg";
-import deleteIcon from "./baseline-delete-24px.svg";
-
-export { refreshIcon, deleteIcon };
